🎨Rename: Update 파일 구조 수정
파일 구조 수정 related to #24
